Cells in grid will be drawn on next page
Resolve Image Anchor Error
Resolve Error about GridLayout where Empty Components doesn't have width in with Span
Change Explict casting to Math.round between int and float
Cells in the grid layout are not cut between pages. If there is a situation to be drawn between pages, it will be drawn on the next page.